+++ tetanes-core/src/mapper/m069_sunsoft_fme7.rs
@@ -144,7 +144,7 @@ impl MemMap for SunsoftFme7 {
 impl Reset for SunsoftFme7 {}
 
 impl Clock for SunsoftFme7 {
-    fn clock(&mut self) -> usize {
+    fn clock(&mut self) -> u64 {
         let cycles = if self.regs.irq_counter_enabled {
             self.regs.irq_counter = self.regs.irq_counter.wrapping_sub(1);
             if self.regs.irq_counter == 0xFFFF && self.regs.irq_enabled {
@@ -271,7 +271,7 @@ impl Audio {
 }
 
 impl Clock for Audio {
-    fn clock(&mut self) -> usize {
+    fn clock(&mut self) -> u64 {
         if self.clock_timer == 0 {
             self.clock_timer = 1;
             for channel in 0..3 {
